CSPD/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
SQLAlchemy==1.4.45
pydantic[email]
envparse==0.2.0
psycopg2==2.9.5
asyncpg==0.27.0
alembic==1.9.0
pytest==7.2.0
pytest-asyncio==0.20.3
httpx==0.23.3
pre-commit==2.21.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
bcrypt==4.0.1